backend/requirements.txt
Flask==2.3.3
flask-cors
pandas==2.1.1
numpy==1.24.3
openpyxl==3.1.2
odfpy==1.4.1
xlrd==2.0.1
dnspython==2.6.1
chardet
secure-smtplib==0.1.1
imapclient==2.3.1
email-validator==2.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dateutil==2.8.2
python-dotenv
selenium==4.38.0
webdriver-manager==4.0.0
bs4
validate_email_address
ddgs
tldextract